Add .luacheckrc
This provides a basic configuration for [LuaCheck][1], allowing you to
find several common errors with the rom files.

The configuration simply declares all of CC's globals, as well as the
custom os table. We ignore several warnings, nameley:

 - Anything to do with formatting
 - Unused variables - whilst we could rename all of these to _, it seems
   easier to just ignore them for now.
 - Shadowing values - see previous comment.

We also allow declaring globals in the rom/apis folder, but nowhere else.

The linter can be run as follows:

    luacheck -q . | less

[1]: https://github.com/mpeterv/luacheck